HADOOP-10739. Renaming a file into a directory containing the same filename results in a confusing I/O error. Contributed by chang li
git-svn-id: https://svn.apache.org/repos/asf/hadoop/common/branches/branch-2@1606268 13f79535-47bb-0310-9956-ffa450edef68

+++ hadoop-common-project/hadoop-common/src/main/java/org/apache/hadoop/fs/shell/MoveCommands.java
@@ -104,6 +104,9 @@ class MoveCommands {
         throw new PathIOException(src.toString(),
             "Does not match target filesystem");
       }
+      if (target.exists) {
+        throw new PathExistsException(target.toString());
+      }
       if (!target.fs.rename(src.path, target.path)) {
         // we have no way to know the actual error...
         throw new PathIOException(src.toString());
